+++ GuaranteedRate.Sextant.Integration.Tests/Logging/Elasticsearch/ElasticsearchLogAppenderIntegrationTests.cs
@@ -15,7 +15,8 @@ namespace GuaranteedRate.Sextant.Integration.Tests.Logging.Elasticsearch
             {
                 var fields = new Dictionary<string, string>
                 {
-                    {"Application", "Encompass"},
+                    {"Application", "SextantLogger"},
+                    {"Message", "The original team behind the game Portal called the game Narbuncular Drop because the word Narbuncular is unique.  It would make a great test logging message too." },
                     {"Company", "Guaranteed Rate"},
                     {"loggerName", "Guaranteed Rate Encompass Logger"}
                 };